backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
django-filter==23.5
psycopg2-binary==2.9.9
python-decouple==3.8
Pillow==10.2.0
boto3==1.34.34
django-storages==1.14.2
python-magic==0.4.27
mypy==1.8.0
django-stubs==4.2.7
black==24.1.1
isort==5.13.2
flake8==7.0.0